requirements.txt
numpy==1.26.4
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
torchdiffeq
diffusers==0.32.0
plyfile
ipython
matplotlib
Pillow
imageio
imageio-ffmpeg
rembg[gpu]
lpips
einops
safetensors
huggingface_hub
safetensors
sentencepiece
accelerate
transformers==4.49.0
kiui
omegaconf
pyarrow
igraph
h5py
debugpy
kornia
pymeshfix
torch_geometric
pyvista
xatlas
easydict
gsplat
e3nn
trimesh
jaxtyping
redis
loguru
pandas
image-reward
opencv-python
opencv_python_headless
warp_lang==0.10.1
taichi==1.5.0
git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/sam2.git
git+https://github.com/ruihong04/pyTorchChamferDistance
tqdm
loralib
fairscale
PyYAML
peft==0.9.0
einops==0.8.1
xformers==0.0.31
timm==0.6.13